Create class container component: created class container component for user.
